fix loader fall
