--- scopes/react/component-highlighter/bubble/index.ts
@@ -1,3 +1,0 @@
-import styles from './bubble.module.scss';
-
-export const bubble = styles.bubble;
