--- scopes/react/component-highlighter/element-highlighter/new-label/index.ts
@@ -1,2 +1,0 @@
-export { NewLabel } from './new-label';
-export type { NewLabelProps } from './new-label';
