--- scopes/react/component-highlighter/label/component-label/duo-component-bubble.module.scss
@@ -1,24 +1,0 @@
-.duoComponentBubble {
-  display: flex;
-
-  //reset styles from card:
-  padding: 0;
-  border: none;
-  background: none;
-
-  > :first-child {
-    border-top-left-radius: 4px;
-    border-bottom-left-radius: 4px;
-
-    margin-right: 2px;
-  }
-
-  > :last-child {
-    border-top-right-radius: 4px;
-    border-bottom-right-radius: 4px;
-  }
-}
-
-.componentName {
-  display: flex;
-}
