--- scopes/react/component-highlighter/label/component-label/index.ts
@@ -1,2 +1,0 @@
-export { ComponentLabel } from './component-label';
-export type { ComponentLabelProps } from './component-label';
