--- scopes/react/component-highlighter/label/default-label/default-label.module.scss
@@ -1,12 +1,0 @@
-.container {
-  display: flex;
-
-  //reset styles from card:
-  padding: 0;
-  border: none;
-  background: none;
-
-  > * {
-    border-radius: 4px;
-  }
-}
