--- scopes/react/component-highlighter/label/default-label/index.ts
@@ -1,1 +1,0 @@
-export { DefaultLabel } from './default-label';
