+++ scopes/react/component-highlighter/label/label.module.scss
