+++ scopes/react/component-highlighter/label/other-components.tsx
